# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NovaRewards API",
    "version": "1.0.0",
    "description": "REST API for the Nova Rewards blockchain loyalty platform. User endpoints require a Bearer JWT from `POST /auth/login`. Merchant endpoints require an `x-api-key` header.",
    "contact": {
      "name": "Nova Rewards Team"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development"
    },
    {
      "url": "https://api.novarewards.io/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /auth/login`."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Merchant API key returned on registration (shown once)."
      }
    },
...
  },
  "paths": {
...
    "/transactions/record": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Verify and record a Stellar transaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "txHash",
                  "txType"
                ],
                "properties": {
                  "txHash": {
                    "type": "string"
                  },
                  "txType": {
                    "type": "string",
                    "enum": [
                      "distribution",
                      "redemption",
                      "transfer"
                    ]
                  },
                  "amount": {
                    "type": "number"
                  },
                  "fromWallet": {
                    "type": "string"
                  },
                  "toWallet": {
                    "type": "string"
                  },
                  "merchantId": {
                    "type": "integer"
                  },
                  "campaignId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction recorded"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Duplicate transaction"
          }
        }
      }
    },
    "/transactions/merchant-totals": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Aggregate totals for authenticated merchant",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Merchant totals"
          }
        }
      }
    },
    "/transactions/merchant/history": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Paginated transaction history for merchant",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "txType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history"
          }
        }
      }
    },
    "/transactions/report": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Generate merchant transaction report",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report data"
          }
        }
      }
    },
    "/transactions/refund": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Refund a transaction",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transactionId"
                ],
                "properties": {
                  "transactionId": {
                    "type": "integer"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Refund created"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    },
    "/transactions/reconcile": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Reconcile merchant transactions against Horizon",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reconciliation result"
          }
        }
      }
    },
    "/transactions/user/history": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Paginated transaction history (all users)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "walletAddress",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history"
          }
        }
      }
    },
    "/transactions/{walletAddress}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "NOVA payment history for a wallet (Horizon with DB fallback)",
        "parameters": [
          {
            "in": "path",
            "name": "walletAddress",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment history"
          },
          "400": {
            "description": "Invalid wallet address"
          }
        }
      }
    },
...
  }
}